Add wayland code. Uses title as winId as placeholder.
Wayland's internalId is now deprecated, so we can't use it.
KDecoration doesn't expose the Wayland uuid() so we can't use that.

I'll likely enable the appmenu when the wayland window becomes active.
I'll then check if the KDecoration is active too. If it is, then I'll
enable the appmenu and setWinId(uuid). The use will need to focus the
window first before the menu is displayed, but at least it should
work.